requirements.txt
scikit_image==0.13.0
numpy==1.14.0
tensorflow_gpu==1.3.0
scipy==1.0.0
matplotlib==2.0.2
h5py==2.7.0
tqdm==4.19.5
skimage==0.0
scikit_learn==0.19.1
